% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromeres.R
\docType{data}
\name{centromeres}
\alias{centromeres}
\title{Centromeres for hg19.}
\format{An object of class \code{data.frame} with 48 rows and 5 columns.}
\usage{
centromeres
}
\description{
A dataset containing the positions of hg19 centromeres
}
\keyword{datasets}
