% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{compute_responsibilities}
\alias{compute_responsibilities}
\title{Computes responsibilities for a gaussian mixture model with specified
parameters}
\usage{
compute_responsibilities(data, means, variances)
}
\arguments{
\item{data}{(numeric) vector of values to compute probailities for}

\item{means}{(numeric) vector of gaussian means, one per component}

\item{variances}{(numeric) vector of gaussian variances, either length one
or equal to the length of means}
}
\value{
matrix of values with row count equal to the length of data and
column count equal to the length of the means
}
\description{
\code{compute_responsibilities} returns a matrix as in parametric_gmm_fit
except values are responsibilities rather than probabilities
}
