% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{cosine_sim}
\alias{cosine_sim}
\title{Compute cosine similarity of two vectors}
\usage{
cosine_sim(vec1, vec2)
}
\arguments{
\item{vec1}{the first vector}

\item{vec2}{the second vector}
}
\value{
single numeric value for cosine similarity
}
\description{
\code{cosine_sim} returns the cosine similarity of two vectors
}
