% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gmm_em_fixed_var}
\alias{gmm_em_fixed_var}
\title{Optimizes mean differential for a special case of gaussian mixture models}
\usage{
gmm_em_fixed_var(data, means, var)
}
\arguments{
\item{data}{(numeric) vector of values to for the GMM to be fit to}

\item{means}{(numeric) vector of means for GMM components}

\item{var}{(numeric) vector of gaussian variances, either length one
or equal to the length of means}
}
\value{
a single numeric value representing the difference in means
}
\description{
(GMMs) where component means differ by a constant factor
\code{gmm_em_fixed_var} returns a difference optimized using EM and a grid
search
}
