% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gmm_em_only_var}
\alias{gmm_em_only_var}
\title{Determines variance for a gaussian mixture model (GMM) given fixed means 
using a variation of expectation-maximization for the special case of all
equal variances for all components}
\usage{
gmm_em_only_var(data, means, initial_var)
}
\arguments{
\item{data}{(numeric) vector of values to for the GMM to be fit to}

\item{means}{(numeric) vector of means for GMM components}

\item{var}{(numeric) single value, initial estimate for variance}
}
\value{
a single numeric value representing the difference in means
}
\description{
\code{gmm_em_only_var} returns a value of standard deviation for the given
GMM
}
