% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gmm_mad}
\alias{gmm_mad}
\title{Computes weighted median absolute deviations (MADs) for each component of a
gaussian mixture model}
\usage{
gmm_mad(data, means, variances)
}
\arguments{
\item{data}{(numeric) vector of values to compute the MAD over}

\item{means}{(numeric) vector of means for the mixture moel}

\item{variances}{(numeric) vector of gaussian variances, either length one
or equal to the length of means}
}
\value{
a single numeric of the MAD for the model
}
\description{
\code{gmm_mad} returns the weighted median absolute deviation of data fit to
the specified mixture model
}
\examples{
gmm_mad(0:5, c(1, 5), 2)
}
