% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{load_ploidetect_data}
\alias{load_ploidetect_data}
\title{Load data from Ploidetect's output}
\usage{
load_ploidetect_data(path)
}
\arguments{
\item{path}{the path to the ploidetect outputs}
}
\value{
a named list oif data.tables
}
\description{
\code{load_ploidetect_data} takes the path to the ploidetect output and returns a named \code{list} containing \code{cna},
a data.table of the bin-level copy number data, \code{segments}, a data.table of the 
segment-level copy number data, and \code{raw}, a data.table of the raw data.
}
