% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{maf_gmm_fit}
\alias{maf_gmm_fit}
\title{Fits a gaussian mixture-of-alleles model}
\usage{
maf_gmm_fit(depth_data, vaf_data, chr_vec, means, variances, maf_variances,
  maxpeak, lowest = NA, ploidy = NA)
}
\arguments{
\item{depth_data}{input read depth data for all bins}

\item{vaf_data}{input vaf data for all bins}

\item{chr_vec}{vector of chromosomes for all bins}

\item{means}{vector of mixture model means}

\item{variances}{standard deviation for the mixture model}

\item{lowest}{optional unless ploidy specified; what is the copy number of 
the first element of means?}

\item{ploidy}{optional unless lowest specified; what is the most common 
copy number state?}

\item{vaf_variances}{standard deviation for VAF values}
}
\value{
a matrix of likelihoods, where columns are copy number states and 
rows are input observations
}
\description{
\code{maf_gmm_fit} fits a gaussian mixture model for each possible
copy number state to VAF data, computes the probability of each copy number
and returns a matrix of probabilities for each copy number state
}
