% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{map_noisy_segments}
\alias{map_noisy_segments}
\title{Map all segments within an interval to the dominant CN state}
\usage{
map_noisy_segments(cna, chrom, start_pos, end_pos)
}
\arguments{
\item{cna}{a data.table containing the bin-level copy number calls}

\item{chrom}{the chromosome of interest}

\item{start_pos}{the start position of the mapping}

\item{end_pos}{the end position of the mapping - the segments of interest must have end <= this number}
}
\value{
a named list
}
\description{
\code{map_noisy_segments} takes in a genomic region on chromosome \code{chr} between \code{start_pos}
and \code{end_pos} and produces a named list of vectors to be passed to \code{curate_segments} 
indicating a mapping between each segment and the most common segment in the region
}
