% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{match_kde_height}
\alias{match_kde_height}
\title{Estimate variance by comparison to a density estimate}
\usage{
match_kde_height(data, means, sd, comparison_point = NA)
}
\arguments{
\item{data}{input data to estimate density and mix the mixture model to}

\item{means}{the means of the GMM}

\item{sd}{the initial estimate of standard deviation}

\item{comparison_point}{value to compute density function and density 
estimator height for optimization}
}
\value{
an estimation of standard deviation
}
\description{
\code{match_kde_height} returns standard deviation such that the probability
density function of the gaussian mixture model (GMM) specified by the 
"means" parameter closely matches the kernel density estimation for the 
input data
}
