% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{mixpdf}
\alias{mixpdf}
\title{Generate the probability density function (pdf) of the specified gaussian
mixture model (GMM)}
\usage{
mixpdf(x, means, proportions, sd)
}
\arguments{
\item{x}{(numeric) vector of values over which to compute the pdf}

\item{means}{(numeric) vector of means for GMM components}

\item{proportions}{(numeric) vector of GMM component weights}

\item{sd}{(numeric) vector or single value of GMM component variance}
}
\value{
a single numeric value representing the difference in means
}
\description{
\code{mixpdf} returns a vector of probability density over the values of x
}
