% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{mixpdf_function}
\alias{mixpdf_function}
\title{Generate a function that outputs the probability density function of a GMM
given an input range}
\usage{
mixpdf_function(means, proportions, sd)
}
\arguments{
\item{means}{(numeric) vector of means for GMM components}

\item{proportions}{(numeric) vector of GMM component weights}

\item{sd}{(numeric) vector or single value of GMM component variance}
}
\value{
a single numeric value representing the difference in means
}
\description{
\code{mixpdf_function} returns a vector of probability density over the values of x
}
