% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{n50_segments}
\alias{n50_segments}
\title{Compute the n50 length of vector of lengths}
\usage{
n50_segments(lengths)
}
\arguments{
\item{lengths}{input vector of lengths}
}
\value{
the n50 value
}
\description{
\code{n50_segments} returns the n50 length of input lengths
}
