% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{norm_lik}
\alias{norm_lik}
\title{Likelihood function of a gaussian distribution}
\usage{
norm_lik(all_obs, mean, sd)
}
\arguments{
\item{all_obs}{(numeric) vector of values over which to compute the 
likelihood}

\item{mean}{(numeric) mean of the gaussian}

\item{sd}{(numeric) standard deviation of the gaussian}
}
\value{
(numeric) vector of values equal in length to x
}
\description{
\code{norm_lik} returns the value of \eqn{L(x | \mu, \sigma)}, the
likelihood function of a value given a gaussian distribution
}
