% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{normpdf}
\alias{normpdf}
\title{Probability density function of a gaussian distribution}
\usage{
normpdf(x, mean, sd)
}
\arguments{
\item{x}{(numeric) vector of values over which to compute the PDF}

\item{mean}{(numeric) mean of the gaussian}

\item{sd}{(numeric) standard deviation of the gaussian}
}
\value{
(numeric) vector of values equal in length to x
}
\description{
\code{normpdf} returns the value of \eqn{f(x, \mu, \sigma^2)}, the
probability density function (PDF) of a gaussian distribution
}
