% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{parametric_gmm_fit}
\alias{parametric_gmm_fit}
\title{Fits a gaussian mixture model with specified parameters to the data agnostic
to weights
\code{parametric_gmm_fit} returns a matrix of standardized normal 
probabilities with rows as observations and columns as mixture components}
\usage{
parametric_gmm_fit(data, means, variances)
}
\arguments{
\item{data}{(numeric) vector of values to compute probailities for}

\item{means}{(numeric) vector of gaussian means, one per component}

\item{variances}{(numeric) vector of gaussian variances, either length one
or equal to the length of means}
}
\value{
matrix of values with row count equal to the length of data and
column count equal to the length of the means
}
\description{
Fits a gaussian mixture model with specified parameters to the data agnostic
to weights
\code{parametric_gmm_fit} returns a matrix of standardized normal 
probabilities with rows as observations and columns as mixture components
}
\examples{
parametric_gmm_fit(0:5, c(1, 5), 2)
}
