% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{props}
\alias{props}
\title{Convert vector into probability distribution}
\usage{
props(x)
}
\arguments{
\item{x}{(numeric) input vector}
}
\value{
a numeric vector of equal length to input
}
\description{
\code{props} returns the proportion-transformed version of input vector
}
