% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{rms}
\alias{rms}
\title{Compute the root-mean-squared (rms) deviation of a vector}
\usage{
rms(x, na.rm = T)
}
\arguments{
\item{x}{(numeric) input vector}

\item{na.rm}{(boolean) should NAs be removed?}
}
\value{
the rmsd of the input
}
\description{
\code{rms} returns the rms of the input vector
}
