% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{table_vec}
\alias{table_vec}
\title{Return a vectorized counts table}
\usage{
table_vec(x)
}
\arguments{
\item{x}{the input vector to count unique occurrances of}
}
\value{
a numeric vector of counts, where names are the counted observation
and values are counts
}
\description{
\code{table_vec} returns a vector of counts of unique observations of input
}
