% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{tp_diffploidy}
\alias{tp_diffploidy}
\title{Compute tumour purity at a different ploidy}
\usage{
tp_diffploidy(cov_char, new_ploidy)
}
\arguments{
\item{cov_char}{coverage chracteristics, as returned by 
get_coverage_characteristics}

\item{new_ploidy}{new ploidy to compute tumour purity for}
}
\value{
single numeric value for tumour purity
}
\description{
\code{tp_diffploidy} returns the tumour purity at the given ploidy given 
input coverage characteristics, as returned by get_coverage_characteristics
}
