% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{zt}
\alias{zt}
\title{Z-score}
\usage{
zt(obs, mean, sd, absval = T)
}
\arguments{
\item{obs}{(numeric) value or vector of values}

\item{mean}{(numeric) the mean that should be used for calculation of z}

\item{sd}{(numeric) standard deviation for calculation of z}

\item{absval}{(boolean) should the absolute value of the z-score be returned?}
}
\value{
(numeric) a vector of equal length to the input
}
\description{
\code{zt} returns the z-score (number of standard deviations from the mean)
of the input given a mean and variance
}
\examples{
zt(1, 1, 1)
zt(0, 1, 1)
zt(10, 1, 1)
zt(1:5, 1, 0.3)
}
