% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{zt_p}
\alias{zt_p}
\title{Standardized normal p-value}
\usage{
zt_p(obs, mean, sd)
}
\arguments{
\item{obs}{(numeric) value or vector of values}

\item{mean}{(numeric) the mean of the gaussian being compared against}

\item{sd}{(numeric) standard deviation of the gaussian being compared
against}
}
\value{
(numeric) a p-value between zero and ~0.399
}
\description{
\code{zt_p} returns \eqn{p(x)} by mapping x to a normal distribution with mean
equal to zero and standard deviation equal to one using its z-score
}
\examples{
zt_p(1, 1, 1)
zt_p(0, 1, 1)
zt_p(1:5, 1, 1)
}
