% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_ploidetect}
\alias{plot_ploidetect}
\title{Produce ploidetect-style plots for a copy number profile}
\usage{
plot_ploidetect(
  cnv_data,
  cn_positions,
  cytobands,
  mode = "all",
  seg_lines = F,
  colors = "cnv"
)
}
\arguments{
\item{seg_lines}{a boolean of whether to include vertical dashed lines at each segment breakpoint}

\item{cna}{a data.table containing the regions of interest. Single chromosome only}
}
\value{
a ggplot2 plot
}
\description{
\code{plot_ploidetect} takes a data.table containing the bin-level copy number data for a
locus of interest and an optional argument of whether to indicate segment breakpoints. It returns a 
\code{ggplot2} plot of the data
}
